# M10/tms/tms-api/contract/contract-types-orval/model/vehicleDocumentListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * TMS API
 * Transportation Management System API
 * OpenAPI spec version: 1.0.0
 */
import type { VehicleDocument } from './vehicleDocument';

export interface VehicleDocumentListResponse {
  data: VehicleDocument[];
}